+++ CHANGELOG.md
@@ -1,3 +1,9 @@
+[OST PHP SDK v2.2.3](https://github.com/ostdotcom/ost-sdk-php/tree/v2.2.3)
+---
+
+* Added redemptions module to call redemptions management OST APIs.
+* Added redeemable sku module to call redeemable sku management OST APIs.
+
 [OST PHP SDK v2.2.2](https://github.com/ostdotcom/ost-sdk-php/tree/v2.2.2)
 ---